flask_server/requirements.txt
flask
flask-sqlalchemy
python-dotenv
kivy
flask_restx
pytest